requirements.txt
tensorflow==1.13.1
numpy==1.17.4
Flask==1.1.1
connexion==2.4.0
Keras==2.3.1
dataclasses==0.6
google_api_python_client==1.7.11
Pillow==6.2.1
